fix send http request action & other cleanup
